requirements.txt
datasets==2.18.0
jieba==0.42.1
nltk==3.8.1
numpy==1.26.4
ollama==0.2.0
pandas==2.2.2
py2neo==2021.2.4
pyahocorasick==2.1.0
rouge_chinese==1.0.3
ruamel.base==1.0.0
scikit_learn==1.4.1.post1
seqeval==1.2.2
streamlit==1.32.2
tqdm==4.66.2
transformers==4.39.0
typer==0.12.3
zhipuai==2.0.1
torch==2.2.1+cu118
peft==0.10.0
